requirements.txt
PyNomadBBO==4.4.0
py_wake==2.5.0
shapely==2.0.4
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
geopandas==0.14.4
pyshp==2.3.1
xarray==2024.5.0